Mais exercícios propostos

+++ ExerciciosSenai/src/AtiRevTop/Ati05.java
@@ -4,12 +4,12 @@ import javax.swing.JOptionPane;
 
 public class Ati05 {
 
-	static int valor1, valor2, div, rest;
+	static double valor1, valor2, div, rest;
 	
 	public static void main(String[] args) {
 		
-		valor1 = Integer.parseInt(JOptionPane.showInputDialog("Digite o primeiro valor"));
-		valor2 = Integer.parseInt(JOptionPane.showInputDialog("Digite o segundo valor"));
+		valor1 = Double.parseDouble(JOptionPane.showInputDialog("Digite o primeiro valor"));
+		valor2 = Double.parseDouble(JOptionPane.showInputDialog("Digite o segundo valor"));
 		
 		div = valor1 / valor2;
 		rest = valor1 % valor2;
